requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiogram==3.4.1
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1
readability-lxml==0.8.1
trafilatura==1.7.0
APScheduler==3.10.4
loguru==0.7.2
python-dotenv==1.0.1
prometheus-client==0.20.0
